pyproject.toml
[project]
name = "tabbycat-api2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "python-dotenv>=1.0.1",
]

[tool.pytest.ini_options]
pythonpath = "src"
addopts = [
    "--import-mode=importlib",
]
testpaths = [
    "tests",
]